backend/pyproject.toml
[project]
name = "uipath-onpremise-mcp-server"
version = "0.1.0"
description = "UiPath Onpremise MCP Server with multiple endpoints"
authors = [{ name = "Charles Kim", email = "hyungsoo.kim@uipath.com" }]
dependencies = [
    "uipath>=2.1.54",
    "mcp>=1.11.0",
    "starlette>=0.41.0",
    "uvicorn>=0.32.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.11.0",
    "sse-starlette>=2.1.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "httpx>=0.28.0",
]
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.28.0",
]
